Move the services folder
