+++ client/app/api/protected/route.ts
@@ -1,6 +1,6 @@
 import { getAccessToken, withApiAuthRequired } from "@auth0/nextjs-auth0";
 import { NextResponse } from "next/server";
-import { getApiClient } from "@/app/services";
+import { getApiClient } from "@/services";
 
 export const GET = withApiAuthRequired(async function protected_(req) {
   try {
