+++ client/services/addAuth.ts
