+++ client/services/getApiClient.ts
@@ -1,5 +1,5 @@
 import { addAuth, addBaseUrl } from ".";
-import { paths } from "../../server-client";
+import { paths } from "../server-client";
 import { Fetcher } from "openapi-typescript-fetch";
 
 export const getApiClient = (accessToken?: string) => {
